Move `_load_defn` from `generate` to `__main__`
